Version 0.3
Adjusted the player character movement logic so that it no longer uses a hard coded vector to constantly move towards. It now uses the players forward vector.

Due to the above change, a bug in the animation controller manifested. The animation speed would be slowed down as the animation blendspace was still only using the players x input vector, so when the player is rotated to face and run in the y direction, the animation speed would not adjust. This has also been fixed.