Version 0.5
Added a UI manager class (HUD class) and partially implemented the main menu, alongside a simple pause menu.

Created 2 interfaces, one that allows the world to communicate with the player character and one that allows communication between the player HUD class and the world (including player).

Created a temp starting area and spawned the player next to a temp camera thats used for the main menu. Once play is pressed, the player moves to its starting location (0x, 0y) and the camera is switched back to the players view.

The transition speeds of the cameras has been correctly adjusted.