version 0.7
Fully implemented a modular save system with the ability to save game states whenever.

Currently the game automatically creates a new save game and saves it in a save slot if one doesnt already exists. Else it will load the save game from a slot.

Added the ability to reset progress aswell with a simple reset progress button in the main menu.

Clicking 'quit' or 'main menu' button in the pause menu will save the game, then either end the game session or return the player to the main menu.