Version 0.8
To fix the map data structure issue, i created a function that manually removes the given item in the map, and adds a new one with the updated structure.

The upgrade system now works.

A bug was observed when making more than 4 upgrade level purchases... the buy button needs to become inactive once we reach max upgrade level.

The upgrade system needs to be implemented for stamina next.