Version 0.9.1
Fixed the most annoying bug ever. On restarting my game from c++, the engine decided to delete one of the variables inside of my savegame blueprint and reset the same variable inside of my player character.

Took me a while to figure this out but its all now fixed and working as intended.