Version 0.9.2
Changed the way we save our coins. Now we hold a 'TotalCoins' variable inside of the save game bp and add all of the coins we accumulate throughout each run (this means for each run, our coins counter is set to 0 and will only track the amount of coins we earn on each run).

The total coins will be displayed in the shop.

Made several modifications to the main menu and save systems to accommodate these changes.

Everything seems to work as intended so far.