Version 0.9.4
Tweaked the starting part of the game, the player character no longer used the nav mesh system and just uses a custom bp function to move to the starting position.

Moved the tile spawning logic into a tile spawner bpc and set it up back to what it was before. Now the gamemode has the tile spawner bpc. This way, the logic can easily be transfered to a different part of the code if needs be.