Version 1.0
Added the ability to slide, comes with an animation from mixamo. Fully operational, just need to change the button press from a debug key and handle it properly.

There is also a slight problem with the character capsule collision switching from standing to crouching and making the camera and character jitter.